Exceptions, Lexer and Interpeter changes.
Changes:
 - Implemented print function.
 - Implemented get function.
 - Implemented data() function.
 - Implemented time() function.
 - Implemented sets.
 - Implemented comments.
 - Implemented escape characters support for strings.
 - Implemented index expressions
 - Implemented assignment expressions support for index expressions.
 - Added IndexOutOfRangeException.
 - Added InvalidSetSyntaxException.
 - Added InvalidCallableException.
 - Added ArgumentsAmountException.

+++ src/main/java/indy/pseudokod/ast/NumericLiteral.java
@@ -5,7 +5,7 @@ public class NumericLiteral extends Expression {
 
     public NumericLiteral(String value) {
         super(NodeType.NumericLiteral);
-        this.value = Long.parseLong(value);
+        this.value = Double.parseDouble(value);
     }
 
     public double value() {
